ASoC: SOF: hsw: remove .get_reply()
The .get_reply() callback is always called by the IPC core
immediately upon a wake-up from the platform IPC drived's interrupt
handler. Therefore it can be directly called from the interrupt
handler, avoiding the indirection. To make this possible the message
has to be accessible from the "struct snd_sof_dev" instance and the
.ipc_lock spin-lock has to be taken while manipulating the message.

Signed-off-by: Guennadi Liakhovetski <guennadi.liakhovetski@linux.intel.com>