ASoC: SOF: Introduce descriptors for SOF client
Introduce descriptors for SOF client driver and SOF client
device along with APIs for registering and unregistering
a SOF client driver, sending IPCs from the client device
and accessing the SOF core debugfs root entry.

Along with this, add a couple of new
members to struct snd_sof_dev that will be used for
maintaining the list of clients.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>
Signed-off-by: Fred Oh <fred.oh@linux.intel.com>