ASoC: SOF: ops: Add new op for client registration
Add a new op for registering clients. The clients to be
registered depends on the capabilities of each platform,
some of which might depend on the ACPI/DT information
and others could be advertised by the DSP in the FW_READY
IPC message after the FW boot is complete.

For now, this patch only adds the dummy IPC test client for
all Intel platforms.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>