virtual_bus: Add uevent callback
The SOF audio driver uses the virtual_bus implementation
for creating virtbus devices that act as client devices
for the SOF core device.

The virtbus driver for each client is separated into different
modules. Without the uevent callback to set the MODALIAS for
the virtbus device, the virtbus device never really gets probed.

This patch is based on:
http://patchwork.ozlabs.org/patch/1247882/

Suggested-by: Pierre-Louis Bossart <pierre-louis.bossart@linux.intel.com>
Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>