virtual-bus: Add generic PM ops for virtual_bus_type
When runtime PM is enabled for the virtbus devices but
the runtime PM callbacks for the device driver are not set,
the device fails to resume with -ENOSYS error.
This happens for devices that don't need to do anything
special in their runtime PM callbacks but need to be
resumed in order for their parents to get resumed.

Add generic PM ops for the virtual_bus_type in order to avoid
adding runtime PM callbacks that do nothing but return 0 to
avoid the above error.

Signed-off-by: Ranjani Sridharan <ranjani.sridharan@linux.intel.com>