Implementation of Ancillary Bus
This is the initial implementation of the Ancillary Bus,
ancillary_device and ancillary_driver.  The ancillary bus is
a software based bus intended to support registering
ancillary_devices and ancillary_drivers and provide matching
between them and probing of the registered drivers.

The bus will support probe/remove shutdown and
suspend/resume callbacks.

Change-Id: Ifdb98949c23119622eb9db66e1bf7ee281fb9a3e
Signed-off-by: Dave Ertman <david.m.ertman@intel.com>
Signed-off-by: Kiran Patil <kiran.patil@intel.com>