ASoC: SOF: parse cavs config data for clock config
The default clock configuration in CAVS platform can be extracted from
config data in the extended manifest. New member, clk_config, is added to
struct snd_sof_dev to store the FW clock config information and that this
will be used later to perform platform-specific operations
in the post_fw_run op. This clk_config will be initialized with
negative number not to break backward compatibility.

Signed-off-by: Fred Oh <fred.oh@linux.intel.com>